backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
stripe
requests
pydantic
pydantic-settings
PyJWT
cryptography
httpx
logtail-python
prometheus-client
redis
fastapi-limiter
sentry-sdk[fastapi]